Update Consolidated CDA Templates for Clinical Notes (US Realm) DSTU R2.1.sch
When we check for R1.1 templateIds, we currently check against a specific set of R2.1 templateId root + extension combos and then make sure that there's a matching R1.1 templateId.

But this doesn't seem particularly robust against future changes. Suppose we bump the extension for one of these templates; we'd need to either have two copies of it (one with the old extension and one with the new) or else we'd lose the ability to make sure that the old root+extension combo still had a matching R1.1 templateId!

I think the better thing to do is to say "in a C-CDA document, an R1.1 templateId with ANY extension at all needs to have a matching extensionless R1.1 templateId." That solves the versioning problem, simplifies the logic, and feels more or less like what C-CDA's trying to go for anyway!